use the same method to configure the urls
